fix: preserve unlimited tier when addon adds max_volumes
Use CASE WHEN to ensure max_volumes=0 (unlimited) is not overridden
by addon's extra_max_volumes. Cast result to bigint for consistent
sqlc type generation.

+++ packages/db/pkg/testutils/queries/models.go
@@ -218,7 +218,7 @@ type TeamLimit struct {
 	MaxVcpu                  int32
 	MaxRamMb                 int32
 	DiskMb                   int32
-	MaxVolumes               int32
+	MaxVolumes               int64
 }
 
 type Tier struct {
